make groundtest and unitest peers of rocketchip, with their own packages, harnesses and configs
