+++ src/main/scala/util/ConfigUtils.scala
@@ -1,0 +1,11 @@
+// See LICENSE for license details.
+
+package util
+
+import scala.math.max
+
+object ConfigUtils {
+  def max_int(values: Int*): Int = {
+    values.reduce((a, b) => max(a, b))
+  }
+}
